Reorganized modules that interact with the database. Improved structure. Cleaned up manage.py script.
