+++ acousticbrainz/views/index.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template
-from acousticbrainz.data import get_last_submitted_tracks, get_stats
+from acousticbrainz.data.stats import get_last_submitted_tracks, get_stats
 
 index_bp = Blueprint('index', __name__)
